android/app/src/main/java/id/ac/ugm/sv/trpl/glucosemonitoring/domain/usecase/DeleteEventUseCase.kt
package id.ac.ugm.sv.trpl.glucosemonitoring.domain.usecase

import id.ac.ugm.sv.trpl.glucosemonitoring.domain.common.wrapper.Result
import id.ac.ugm.sv.trpl.glucosemonitoring.domain.repository.IEventRepository
import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
import io.reactivex.rxjava3.core.Flowable
import io.reactivex.rxjava3.schedulers.Schedulers
import javax.inject.Inject

class DeleteEventUseCase @Inject constructor(
    private val eventRepository: IEventRepository,
) {
    
    operator fun invoke(id: Int): Flowable<Result<Nothing>> {
        return eventRepository.deleteEvent(id)
            .subscribeOn(Schedulers.io())
            .observeOn(AndroidSchedulers.mainThread())
    }
    
}